+++ app/Http/Requests/Api/OfferStoreRequest.php
@@ -52,6 +52,7 @@ class OfferStoreRequest extends FormRequest
             'exam' => 'nullable|string',
             'requirements' => 'nullable|string',
             'niveau' => 'nullable|string',
+            'location' => 'nullable|string',
 
             #HUBOFFERS table
             'sort_flag' => 'nullable|integer',