+++ Changelog.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [UnReleased v0.5]
+
+## 2021-11-10/kw
+## Changed
+- changed relatedOffers synchronization in "saveRelatedData" function in OfferController back to previous functionality 
+- changed relatedOfferRule back to previous state
+
+## 2021-11-08/kw
+## Added
+- "location" to Offer(update/store)request rules 
+- "location" to ExternalOfferUpdaterequest rules 
+## Changed
+- saveRelatedData function to delete rows in pivot tables related to Offer
+- Offer related pivot tables no longer delete data unless explicitly sending delete request with §key =(no value) ;  
+- relatedOfferRule 
+
 ## [Released v0.4.2]
 
 ## 2021-10-14/kw